@@ -5,22 +5,61 @@
     "homepage": "https://jira.mongodb.org/browse/PHPLIB",
     "license": "Apache-2.0",
     "authors": [
-        { "name": "Hannes Magnusson", "email": "bjori@mongodb.com" },
+        { "name": "Andreas Braun", "email": "andreas.braun@mongodb.com" },
         { "name": "Jeremy Mikola", "email": "jmikola@gmail.com" },
-        { "name": "Derick Rethans", "email": "github@derickrethans.nl" }
+        { "name": "Jérôme Tamarelle", "email": "jerome.tamarelle@mongodb.com" }
     ],
     "require": {
-        "php": ">=5.4",
-        "ext-mongodb": "^1.2.0"
+        "php": "^7.4 || ^8.0",
+        "ext-hash": "*",
+        "ext-json": "*",
+        "ext-mongodb": "^1.18.0",
+        "composer-runtime-api": "^2.0",
+        "psr/log": "^1.1.4|^2|^3",
+        "symfony/polyfill-php80": "^1.27",
+        "symfony/polyfill-php81": "^1.27"
     },
     "require-dev": {
-        "phpunit/phpunit": "^4.8"
+        "doctrine/coding-standard": "^12.0",
+        "rector/rector": "^1.1",
+        "squizlabs/php_codesniffer": "^3.7",
+        "symfony/phpunit-bridge": "^5.2",
+        "vimeo/psalm": "^5.13"
     },
     "autoload": {
         "psr-4": { "MongoDB\\": "src/" },
         "files": [ "src/functions.php" ]
     },
     "autoload-dev": {
-        "psr-4": { "MongoDB\\Tests\\": "tests/" }
+        "psr-4": {
+            "MongoDB\\Tests\\": "tests/"
+        },
+        "files": [ "tests/PHPUnit/Functions.php" ]
+    },
+    "scripts": {
+        "bench": "cd benchmark && composer update && vendor/bin/phpbench run --report=aggregate",
+        "checks": [
+            "@check:cs",
+            "@check:psalm",
+            "@check:rector"
+        ],
+        "check:cs": "phpcs",
+        "check:psalm": "psalm",
+        "check:rector": "rector --ansi --dry-run",
+        "fix:cs": "phpcbf",
+        "fix:psalm:baseline": "psalm --set-baseline=psalm-baseline.xml",
+        "fix:rector": "rector process --ansi",
+        "test": "simple-phpunit"
+    },
+    "extra": {
+        "branch-alias": {
+            "dev-master": "1.x-dev"
+        }
+    },
+    "config": {
+        "allow-plugins": {
+            "dealerdirect/phpcodesniffer-composer-installer": true
+        },
+        "sort-packages": true
     }
 }
