@@ -28,102 +28,125 @@
  */
 
 // Basic setup for Roundcube calendar client class
-function rcube_calendar(settings)
-{
-    // extend base class
-    rcube_libcalendaring.call(this, settings);
+function rcube_calendar(settings) {
+  // extend base class
+  rcube_libcalendaring.call(this, settings);
 
-    // create new event from current mail message
-    this.create_from_mail = function(uid)
-    {
-      if (!uid && !(uid = rcmail.get_single_uid())) {
-        return;
-      }
-
-      var url = {_mbox: rcmail.env.mailbox, _uid: uid, _framed: 1},
-        buttons = {},
-        button_classes = ['mainaction save', 'cancel'],
-        title = rcmail.gettext('calendar.createfrommail'),
-        dialog = $('<iframe>').attr({
-            id: 'kolabcalendarinlinegui',
-            name: 'kolabcalendardialog',
-            src: rcmail.url('calendar/dialog-ui', url)
-          });
-
-      // dialog buttons
-      buttons[rcmail.gettext('save')] = function() {
-        var frame = rcmail.get_frame_window('kolabcalendarinlinegui');
-        frame.rcmail.command('event-save');
-      };
-
-      buttons[rcmail.gettext('cancel')] = function() {
-        dialog.dialog('destroy');
-      };
-
-      // open jquery UI dialog
-      window.kolab_event_dialog_element = dialog = rcmail.show_popup_dialog(dialog, title, buttons, {
-          button_classes: button_classes,
-          minWidth: 500,
-          width: 600,
-          height: 600
-      });
-    };
-
-    // handler for attachment-save-calendar commands
-    this.save_to_calendar = function(id)
-    {
-      // TODO: show dialog to select the calendar for importing
-      if ((this.selected_attachment || id) && window.rcube_libcalendaring) {
-        rcmail.http_post('calendar/mailimportattach', {
-            _uid: rcmail.env.uid,
-            _mbox: rcmail.env.mailbox,
-            _part: id || this.selected_attachment,
-            // _calendar: $('#calendar-attachment-saveto').val(),
-          }, rcmail.set_busy(true, 'itip.savingdata'));
-      }
-    };
-}
-
-
-/* calendar plugin initialization (for non-calendar tasks) */
-window.rcmail && rcmail.addEventListener('init', function(evt) {
-  if (rcmail.task != 'calendar') {
-    var cal = new rcube_calendar($.extend(rcmail.env.calendar_settings, rcmail.env.libcal_settings));
-    window.cal = cal;
-    //PAMELA - Add custom events
-    rcmail.register_command("add-event-from-shortcut", function() { cal.create_event_from_somewhere(); });
-    rcmail.enable_command('add-event-from-shortcut', true);
-    // register create-from-mail command to message_commands array
-    if (rcmail.env.task == 'mail') {
-      rcmail.register_command('calendar-create-from-mail', function() { cal.create_from_mail(); });
-      rcmail.register_command('attachment-save-calendar', function(p) { cal.save_to_calendar(p); });
-
-      if (rcmail.env.action != 'show') {
-        rcmail.env.message_commands.push('calendar-create-from-mail');
-      }
-      else {
-        rcmail.enable_command('calendar-create-from-mail', true);
-      }
-
-      rcmail.addEventListener('beforemenu-open', function(p) {
-        if (p.menu == 'attachmentmenu') {
-          cal.selected_attachment = p.id;
-          var mimetype = rcmail.env.attachments[p.id];
-          rcmail.enable_command('attachment-save-calendar', mimetype == 'text/calendar' || mimetype == 'text/x-vcalendar' || mimetype == 'application/ics');
-        }
-      });
+  // create new event from current mail message
+  this.create_from_mail = function (uid) {
+    if (!uid && !(uid = rcmail.get_single_uid())) {
+      return;
     }
 
-    //PAMELA - add custom actions after init
-    rcmail.triggerEvent("plugin.calendar.initialized", cal);
+    var url = { _mbox: rcmail.env.mailbox, _uid: uid, _framed: 1 },
+      buttons = {},
+      button_classes = ['mainaction save', 'cancel'],
+      title = rcmail.gettext('calendar.createfrommail'),
+      dialog = $('<iframe>').attr({
+        id: 'kolabcalendarinlinegui',
+        name: 'kolabcalendardialog',
+        src: rcmail.url('calendar/dialog-ui', url),
+      });
 
-  }
+    // dialog buttons
+    buttons[rcmail.gettext('save')] = function () {
+      var frame = rcmail.get_frame_window('kolabcalendarinlinegui');
+      frame.rcmail.command('event-save');
+    };
 
-  rcmail.register_command('plugin.calendar', function() { rcmail.switch_task('calendar'); }, true);
+    buttons[rcmail.gettext('cancel')] = function () {
+      dialog.dialog('destroy');
+    };
 
-  rcmail.addEventListener('plugin.ping_url', function(p) {
-    var action = p.action;
-    p.action = p.event = null;
-    new Image().src = rcmail.url(action, p);
+    // open jquery UI dialog
+    window.kolab_event_dialog_element = dialog = rcmail.show_popup_dialog(
+      dialog,
+      title,
+      buttons,
+      {
+        button_classes: button_classes,
+        minWidth: 500,
+        width: 600,
+        height: 600,
+      },
+    );
+  };
+
+  // handler for attachment-save-calendar commands
+  this.save_to_calendar = function (id) {
+    // TODO: show dialog to select the calendar for importing
+    if ((this.selected_attachment || id) && window.rcube_libcalendaring) {
+      rcmail.http_post(
+        'calendar/mailimportattach',
+        {
+          _uid: rcmail.env.uid,
+          _mbox: rcmail.env.mailbox,
+          _part: id || this.selected_attachment,
+          // _calendar: $('#calendar-attachment-saveto').val(),
+        },
+        rcmail.set_busy(true, 'itip.savingdata'),
+      );
+    }
+  };
+}
+
+/* calendar plugin initialization (for non-calendar tasks) */
+window.rcmail &&
+  rcmail.addEventListener('init', function (evt) {
+    if (rcmail.task != 'calendar') {
+      var cal = new rcube_calendar(
+        $.extend(rcmail.env.calendar_settings, rcmail.env.libcal_settings),
+      );
+      window.cal = cal;
+      //PAMELA - Add custom events
+      rcmail.register_command('add-event-from-shortcut', function () {
+        cal.create_event_from_somewhere();
+      });
+      rcmail.enable_command('add-event-from-shortcut', true);
+      // register create-from-mail command to message_commands array
+      if (rcmail.env.task == 'mail') {
+        rcmail.register_command('calendar-create-from-mail', function () {
+          cal.create_from_mail();
+        });
+        rcmail.register_command('attachment-save-calendar', function (p) {
+          cal.save_to_calendar(p);
+        });
+
+        if (rcmail.env.action != 'show') {
+          rcmail.env.message_commands.push('calendar-create-from-mail');
+        } else {
+          rcmail.enable_command('calendar-create-from-mail', true);
+        }
+
+        rcmail.addEventListener('beforemenu-open', function (p) {
+          if (p.menu == 'attachmentmenu') {
+            cal.selected_attachment = p.id;
+            var mimetype = rcmail.env.attachments[p.id];
+            rcmail.enable_command(
+              'attachment-save-calendar',
+              mimetype == 'text/calendar' ||
+                mimetype == 'text/x-vcalendar' ||
+                mimetype == 'application/ics',
+            );
+          }
+        });
+      }
+
+      //PAMELA - add custom actions after init
+      rcmail.triggerEvent('plugin.calendar.initialized', cal);
+    }
+
+    rcmail.register_command(
+      'plugin.calendar',
+      function () {
+        rcmail.switch_task('calendar');
+      },
+      true,
+    );
+
+    rcmail.addEventListener('plugin.ping_url', function (p) {
+      var action = p.action;
+      p.action = p.event = null;
+      new Image().src = rcmail.url(action, p);
+    });
   });
-});
