@@ -508,10 +508,10 @@
             }, true);
 
             ChatSetupConfig();
-
             if (rcmail.env.launch_chat_frame_at_startup === true)
             {
                 mel_metapage.Functions.change_frame('rocket', false);
+                mel_metapage.Functions.change_frame('tchap', false);
             }
 
             //Ajustement de la barre des tâches
@@ -828,20 +828,22 @@ $(document).ready(() => {
         });
     } else {
         rcmail.addEventListener("init", async function() {
-            let querry = $("#user-picture");
-            // console.log("ping", _ping, rcmail.env.rocket_chat_url + "/avatar/" + rcmail.env.username);
-            querry.html('<img alt="" src=' + rcmail.env.rocket_chat_url + "avatar/" + rcmail.env.username + ' />');
-            let image = querry.find("img")[0];
-            image.onerror = function() {
-                $("#user-picture").html("<span class='no-image avatar'>" + rcmail.env.username.slice(0, 1).toUpperCase() + "</span>");
-            };
-
-            let querry_big = $("#user-big-picture");
-            querry_big.html('<img alt="" src=' + rcmail.env.rocket_chat_url + "avatar/" + rcmail.env.username + ' />');
-            let image_big = querry_big.find("img")[0];
-            image_big.onerror = function() {
-                $("#user-big-picture").html("<span class='no-image big-avatar'>" + rcmail.env.username.slice(0, 1).toUpperCase() + "</span>");
-            };
+            if (rcmail.env.rocket_chat_url) {
+                let querry = $("#user-picture");
+                // console.log("ping", _ping, rcmail.env.rocket_chat_url + "/avatar/" + rcmail.env.username);
+                querry.html('<img alt="" src=' + rcmail.env.rocket_chat_url + "avatar/" + rcmail.env.username + ' />');
+                let image = querry.find("img")[0];
+                image.onerror = function() {
+                    $("#user-picture").html("<span class='no-image avatar'>" + rcmail.env.username.slice(0, 1).toUpperCase() + "</span>");
+                };
+    
+                let querry_big = $("#user-big-picture");
+                querry_big.html('<img alt="" src=' + rcmail.env.rocket_chat_url + "avatar/" + rcmail.env.username + ' />');
+                let image_big = querry_big.find("img")[0];
+                image_big.onerror = function() {
+                    $("#user-big-picture").html("<span class='no-image big-avatar'>" + rcmail.env.username.slice(0, 1).toUpperCase() + "</span>");
+                };
+            }
         });
 
         // Fermeture du bienvenue
