@@ -198,6 +198,7 @@ class mel_moncompte extends rcube_plugin {
       $this->register_action('plugin.invitation',               array($this,'invitation'));
       $this->register_action('plugin.no_invitation',            array($this,'no_invitation'));
       $this->register_action('plugin.set_default_resource',     array($this,'set_default_resource'));
+      $this->register_action('plugin.ressources_calendar_agendas_tri_alpha', [$this, 'ressources_calendar_agendas_tri_alpha']);
 
       // register actions
       $this->register_action('plugin.mel_resources_bal', array($this,'resources_bal_init'));
@@ -378,6 +379,7 @@ class mel_moncompte extends rcube_plugin {
           $default_calendar = $user->getDefaultCalendar();
           $acl = ($calendar->asRight(LibMelanie\Config\ConfigMelanie::WRITE) ? $this->gettext('read_write') : ($calendar->asRight(LibMelanie\Config\ConfigMelanie::READ) ? $this->gettext('read_only') : ($calendar->asRight(LibMelanie\Config\ConfigMelanie::FREEBUSY) ? $this->gettext('show') : $this->gettext('none'))));
           $shared = $user->uid != $calendar->owner || !$calendar->asRight(LibMelanie\Config\ConfigMelanie::WRITE);
+          $name_editable = ($user->uid === $calendar->owner && $id !== $user->uid) ? true : false;
           $is_default = $default_calendar->id == $calendar->id;
           $this->rc->output->set_env("resource_id", $id);
           $this->rc->output->set_env("resource_name", $shared ? "(" . $calendar->owner . ") " . $calendar->name : $calendar->name);
@@ -386,6 +388,9 @@ class mel_moncompte extends rcube_plugin {
           $this->rc->output->set_env("resource_owner", $calendar->owner);
           $this->rc->output->set_env("resource_default", $default_calendar->id == $calendar->id);
           $this->rc->output->set_env("resource_invitation", !isset($no_invitation[$calendar->id]));
+          $this->rc->output->set_env("resource_color", $this->rc->config->get('color_calendars', null)[$id]);
+          $this->rc->output->set_env("resource_name_editable", $name_editable);
+          $this->rc->output->set_env("resource_showalarms", $this->rc->config->get('alarm_calendars', null)[$id]);
 
           if ($acl === $this->gettext('read_only') || M2calendar::is_external($calendar->id)) {
             $this->rc->output->set_env("show_invitations", false);
@@ -422,6 +427,8 @@ class mel_moncompte extends rcube_plugin {
         }
       }
       else {
+        $account = rcube_utils::get_input_value('_account', rcube_utils::INPUT_GET);
+
         // register UI objects
         $this->rc->output->add_handlers(
             array(
@@ -429,9 +436,25 @@ class mel_moncompte extends rcube_plugin {
                 'mel_resources_type_frame'    => array($this,'mel_resources_type_frame'),
             )
         );
-        $this->rc->output->set_env("account", rcube_utils::get_input_value('_account', rcube_utils::INPUT_GET));
+        $config = [
+                      'command' => 'rcs-agenda-tri-alpha',
+            'class' => 'alpha-sort-icon'.($account === null ? '' : ' disabled'),
+            'innerclass' => 'inner',
+            'id' => 'rcs-agenda-tri-alpha',//tb_label_popup
+            'title' => 'mel_moncompte.sort_title', // gets translated
+            'type' => 'link',
+            'data-type' => 'link',
+            'label' => 'mel_moncompte.sort', // maybe put translated version of "Labels" here?
+            // 'wrapper'    => 'li',
+        ];
+
+        if (isset($account)) $config['disabled'] = 'disabled';
+
+        $this->add_button($config, 'rcs_moncompte_header');
+        $this->rc->output->set_env("account", $account);
         $this->rc->output->set_env("resources_action", "agendas");
         $this->rc->output->include_script('list.js');
+        $this->include_script('events.js');
         $this->rc->output->set_pagetitle($this->gettext('resources'));
         $this->rc->output->send('mel_moncompte.resources_elements');
       }
@@ -444,6 +467,27 @@ class mel_moncompte extends rcube_plugin {
       return false;
     }
   }
+
+  public function ressources_calendar_agendas_tri_alpha() {
+    $user = driver_mel::gi()->getUser();
+    $calendars = $user->getSharedCalendars();
+
+    $calendars = mel_helper::Enumerable($calendars)->orderBy(function ($key, $value) {
+      return ord($value->name);
+    });
+
+    $sort = [];
+    $it = 1;
+    foreach ($calendars as $calendar) {
+      if ($calendar->id !== $user->uid) $sort[(++$it)+($calendar->owner === $user->uid ? 2000 : 3000)] = driver_mel::gi()->mceToRcId($calendar->id);
+    }
+
+    $this->rc->user->save_prefs(['sort_agendas' => $sort]);
+
+    echo 'reload';
+    exit; 
+  }
+
   /**
    * Initialisation du menu ressources pour les Contacts
    * Affichage du template et gestion de la sélection
