@@ -177,7 +177,8 @@ class Moncompte {
 
 		// MANTIS 0005002: [MAA] Modification de mdp pour les BALF
 		$password = rcmail::get_instance()->get_user_password();
-		if ($user->uid != self::get_current_user_name() 
+		if ($user->uid != self::get_current_user_name()
+				&& rcmail::get_instance()->config->get('moncompte_agri_auth_delegated', false)
 				&& $user->load(['is_agriculture']) 
 				&& $user->is_agriculture) {
 			$value = str_replace('AGRI.Lien: uid=', '', $user->liens_import);
