@@ -1,9 +1,9 @@
 /**
-	 * Affiche la première lettre si l'image n'est pas trouvée
-	 * @param {string} iconId Id de l'image
-	 * @param {string} iconId Id de l'overlay si l'image n'est pas chargée
-	 * @param {string} title Titre du lien
-	 */
+   * Affiche la première lettre si l'image n'est pas trouvée
+   * @param {string} iconId Id de l'image
+   * @param {string} iconId Id de l'overlay si l'image n'est pas chargée
+   * @param {string} title Titre du lien
+   */
 function imgError(iconId = null, noImageId = null, title) {
   let iconImage = $('#' + iconId);
   let noImage = $('#' + noImageId);
@@ -15,6 +15,8 @@ function imgError(iconId = null, noImageId = null, title) {
       : null;
 
   iconImage.hide();
-  noImage.html(firstLetter);
+  if (noImage.text() === '') {
+    noImage.html(firstLetter);
+  }
   noImage.css('display', 'flex');
 }
