@@ -1,14 +1,15 @@
 /**
- * @namespace Plugin.Workspace
+ * @namespace Workspace
  * @property {MainWorkspace} Index
  * @property {PartManager} PartManager
  * @property {WspChatManager} WspChatManager
  * @property {Workspace} Workspace
  * @property {module:Planning} Planning
- *
+ * @memberof Plugins
  */
 
 /**
+ * Contient les Classes lié au planning
  * @module Planning
  * @property {PlanningManager} PlanningManager
  * @property {FullCalendarEvent} FullCalendarEvent
