@@ -22,7 +22,7 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-require_once __DIR__ . '/../config.inc.php';
+
 
 /**
  * Classe de gestion des mails
@@ -35,30 +35,30 @@ class Mail
   /**
    * Constructeur privé pour ne pas instancier la classe
    */
-  private function __construct()
-  {
-  }
+  private function __construct() {}
 
   /**
    * Méthode d'envoi du message notification à l'organisateur à la création du sondage
    *
    * @param $attendee informations de l'organisateur
    * @param $attendee informations du participant
-   * @return boolean
+   * @param $appointment informations de l'évènement
    */
 
   public static function SendOrganizerAppointmentMail($organizer, $attendee, $appointment)
   {
-    global $config;
+    require __DIR__ . '/../config.inc.php';
+
     $subject = $config["organizer_mail_subject"];
     $from = $config['mail_from'];
     $to = '=?UTF-8?B?' . base64_encode('"' . $organizer->name . '"') . '?=' . "\r\n <" . $organizer->email . ">";
     $body = file_get_contents(__DIR__ . '/templates/organizer_appointment_mail.html');
+
     // Replace elements
-    $subject = str_replace("%%attendee_name%%", $attendee['name'], $subject);
+    $subject = str_replace("%%attendee_name%%", $attendee['name'] . ' ' . $attendee['firstname'], $subject);
     $subject = str_replace("%%appointment_date_day%%", $appointment['date_day'], $subject);
     $body = str_replace("%%appointment_object%%", $appointment['object'], $body);
-    $body = str_replace("%%attendee_name%%", $attendee['name']. ' ' .$attendee['firstname'], $body);
+    $body = str_replace("%%attendee_name%%", $attendee['name'] . ' ' . $attendee['firstname'], $body);
     $body = str_replace("%%attendee_email%%", $attendee['email'], $body);
     $body = str_replace("%%appointment_date%%", $appointment['date'], $body);
     $body = str_replace("%%appointment_description%%", $appointment['description'], $body);
@@ -72,12 +72,14 @@ class Mail
    *
    * @param $attendee informations de l'organisateur
    * @param $attendee informations du participant
-   * @return boolean
+   * @param $appointment informations de l'évènement
+   * @param $appointment ics de l'évènement
    */
 
   public static function SendAttendeeAppointmentMail($organizer, $attendee, $appointment, $ics)
   {
-    global $config;
+    require __DIR__ . '/../config.inc.php';
+
     $subject = $config["attendee_mail_subject"];
     $from = $config['mail_from'];
     $to = '=?UTF-8?B?' . base64_encode('"' . $attendee['name'] . '"') . '?=' . "\r\n <" . $attendee['email'] . ">";
@@ -122,6 +124,6 @@ class Mail
 
     $body = str_replace("BEGIN:VEVENT", "METHOD:REQUEST\r\nBEGIN:VEVENT", $body);
 
-    return \LibMelanie\Mail\Mail::Send($from, $to, $subject, $body,null, null, null, null, null, null, null, null, $ics);
+    return \LibMelanie\Mail\Mail::Send($from, $to, $subject, $body, null, null, null, null, null, null, null, null, $ics);
   }
 }
